updated to make separate controller.
