+++ conf/app.routes
@@ -1,5 +1,5 @@
 # microservice specific routes
 
-POST        /check                    uk.gov.hmrc.trusts.controllers.TrustsController.checkExistingTrust
+POST        /check                    uk.gov.hmrc.trusts.controllers.CheckTrustController.checkExistingTrust
 
-POST        /register                  uk.gov.hmrc.trusts.controllers.TrustsController.registration
+POST        /register                  uk.gov.hmrc.trusts.controllers.RegisterTrustController.registration
